% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_fns.R
\name{DirFactor}
\alias{DirFactor}
\title{Gibbs sampler for Depedent Dirichlet factor model.}
\usage{
DirFactor(data, hyper, start = NA, save.path = NA, save.obj = c("sigma",
  "Q", "T.aug", "X", "Y", "er", "delta", "phi"), burnin = 0.2, thinning = 5,
  step = 1000, step.disp = 10)
}
\arguments{
\item{data}{Required. A count matrix with species in rows and biological samples in column.}

\item{hyper}{Required. A list of hyper-parameters in the priors.}

\item{start}{A list of starting values of model parameters. Default is \code{NA} and the starting values
will be generated automatically based on the input data. See Details for the required fields of the list.}

\item{save.path}{A string contains the path to save the MCMC results. 
For example, \code{save.path="~/sim"} will save results of the ith iteration to ~/sim_i.rds.
Default is \code{NA} and a temp directory will be assigned.}

\item{save.obj}{A list of model parameters that will be saved, default is all parameters.}

\item{burnin}{A number between 0 and 1. Fraction of burn-in samples. Default is 0.2.}

\item{thinning}{A positive integer. The MCMC results will be saved every \code{thinning} iterations. 
Default is 5.}

\item{step}{A positive integer. The total number of MCMC iterations. Default is 1000.}

\item{step.disp}{A positive integer. A message will report the number of iterations finished 
every \code{step.disp} iterations. Default is 10.}
}
\value{
A list with two fields: \code{running.time} and \code{save.path}. 
\code{running.time} is the total amount of time for finishing the MCMC simulation. 
\code{save.path} is the path to the saved results.
}
\description{
Gibbs sampler for Depedent Dirichlet factor model.
}
\details{
The Dependent Dirichlet process factor model assumes the observed data is distributed
according to a multinomial distrition for each biological sample, conditioning on the probabilities
of species, which is assumed to follow a Dependent Dirichlet processes a priori. 
The model has two major parts. \code{sigma, Q} directly specify the probabilities of species in each
biological samples and \code{X, Y, er, delta, phi} specify the hyper-prior on \code{Q}. \code{T.aug} is
an auxilary parameter and does not have direct interpretation. More details on model and prior 
specification can be found in Ren et. al. (2016).

Users are required to provide a list containing the values of the hyper-parameters. The list must contain
fields as following. 
\itemize{
 \item \code{alpha} and \code{beta}: hyper-parameters in the prior of \code{sigma}. \code{sigma} follows a 
Poisson process on \eqn{(0,1)} with intensity \eqn{\alpha\sigma^{-1-\beta}\cdot(1-\sigma)^{-1/2+\beta}} a priori.
 \item \code{a.er} and \code{b.er}: hyper-parameters in the prior of \code{er}. \code{er} follows an inverse
  gamma distribution \eqn{1/Gamma(a.er,b.er)} a priori.
 \item \code{m}: number of factors the model is assumed to have. Usually a value much smaller than the number
 of biological samples.
 \item \code{nv}: hyper-parameter of the prior of \code{phi}. \code{phi} follows \eqn{Gamma(nv/2,nv/2)} a priori.
 \item \code{a1} and \code{a2}: hyper-parameters in the prior of \code{delta}. The prior of \code{delta[1]} is \eqn{Gamma(a1,1)}
 and the prior of \code{delta[2:m]} is independent \eqn{Gamma(a2,1)}. \code{a2} must not be smaller than 1 in order to
 achieve factor shrinkage Bhattacharya et al. (2011).
}

If the users want to specify the starting values for the model parameters, they can pass a list with fields
\code{sigma, Q, T.aug, X, Y, er, delta, phi} to the function augment \code{start}. Assume there are
\code{n} biological samples and \code{p} species in \code{data}. Each field is specified as following:
\itemize{
  \item \code{sigma}: a vector with \code{p} components and the starting values have to be in \eqn{(0,1)}.
  \item \code{Q}: a matrix with \code{p} rows and \code{n} columns and the starting values of \code{Q} 
  should be positive if the corresponding cell in \code{data} is positive.
  \item \code{T.aug}: a vector with \code{n} positive components. 
  \item \code{X}: a \eqn{m\times p} matrix.
  \item \code{Y}: a \eqn{m\times n} matrix. 
  \item \code{er}: a positive scalar. 
  \item \code{delta}: a positive vector with \code{m} components.
  \item \code{phi}: a \eqn{n\times m} positive matrix. 
}
}
\examples{
my.hyper = list( nv = 3, a.er = 1, b.er = 0.3, a1 = 3, 
                 a2 = 4, m = 10, alpha = 10, beta = 0 )
my.sim = SimDirFactorBlock( 1e6, n = 22, p = 68, m = 3, my.hyper, K = 2 )
DirFactor( my.sim$data[[1]], my.hyper, save.obj = c("Y", "er"), step.disp = 100 )
}

