% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PlotClustering}
\alias{PlotClustering}
\title{Plot posterior probability of pairwise classification based on posterior 
samples of pairwise distances between biological samples.}
\usage{
PlotClustering(all.dist, labels)
}
\arguments{
\item{all.dist}{A list contains posterior samples of pairwise distances
between biological samples. Each element is a distance matrix.}

\item{labels}{A character vector indicates the label for each biological 
sample.}
}
\value{
A ggplot figure object. It illustrates the posterior probability
of two biological samples being clustered together. The rows and columns 
are sorted by labels of biological samples.
}
\description{
Plot posterior probability of pairwise classification based on posterior 
samples of pairwise distances between biological samples.
}

