% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PlotStatis}
\alias{PlotStatis}
\title{Plot ordination results with credible region based on results from DirFactor.}
\usage{
PlotStatis(all.cov, n.dim = 2, labels = NA, types = NA, dist = F)
}
\arguments{
\item{all.cov}{Posterior samples of normalized Gram matrix/distance matrix between biological samples.}

\item{n.dim}{Total number of axes visualized. Default is 2.}

\item{labels}{Labels associated with each biological samples. Default is NA. If not NA, labels will be
annote onto the figure.}

\item{types}{Types of each biological samples. Default is NA. If not NA, the contour corresponds to each
biological samples will be colored based on the type of this sample.}

\item{dist}{A logical flag indicates the types of input. Default is FALSE, meaning each element in \code{all.cov}
is a normalized Gram matrix. If set to TRUE, each element in \code{all.cov} is a distance matrix.}
}
\value{
A list contains all pairwise ordination plots between the first \code{n.dim} axes.
}
\description{
Plot ordination results with credible region based on results from DirFactor.
}
\note{
The STATIS method is implemented by \code{DistatisR} (Abdi et. al. (2005)).
}

