% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SimDirFactorCustom}
\alias{SimDirFactorCustom}
\title{Generate synthetic OTU table from DirFactor model using latent factors provided by users.}
\usage{
SimDirFactorCustom(vcounts, sigma, X, Y, er)
}
\arguments{
\item{vcounts}{Required. A vector of total counts per biological sample. Example: 
\code{vcounts=c(1e5,1e6)}.}

\item{sigma}{A length \code{p} vector. \code{p} is the number of species. 
Global abundance parameter for species.}

\item{X}{A \code{m*p} matrix. \code{m} is the number of factors. Latent 
factors for species.}

\item{Y}{A \code{m*n} matrix. \code{n} is the number of biological samples. 
Latent factors for biological samples.}

\item{er}{A positive scalar. Variance of the pure noise.}
}
\value{
A list with \code{length(vcounts)} elements. The \code{i}th elment is 
an OTU table with total counts per biological sample \code{vcounts[i]}.
}
\description{
Generate synthetic OTU table from DirFactor model using latent factors provided by users.
}

